+++ cypress/e2e/invoices/invoices-manually-create-line.cy.js
@@ -1,12 +1,14 @@
 import Invoices from '../../support/fragments/invoices/invoices';
 import NewInvoice from '../../support/fragments/invoices/newInvoice';
+import NewInvoiceLine from '../../support/fragments/invoices/newInvoiceLine';
 import VendorAddress from '../../support/fragments/invoices/vendorAddress';
 import Organizations from '../../support/fragments/organizations/organizations';
 import TopMenu from '../../support/fragments/topMenu';
 
-describe('ui-invoices: Invoice creation', () => {
+describe('Invoices', () => {
   const invoice = { ...NewInvoice.defaultUiInvoice };
   const vendorPrimaryAddress = { ...VendorAddress.vendorAddress };
+  const invoiceLine = { ...NewInvoiceLine.defaultUiInvoiceLine };
 
   before(() => {
     cy.getAdminToken();
@@ -27,11 +29,12 @@ describe('ui-invoices: Invoice creation', () => {
   });
 
   it(
-    'C2299 Manually Create Invoice (thunderjet)',
+    'C2326 Manually create invoice line (thunderjet)',
     { tags: ['smoke', 'thunderjet', 'shiftLeft'] },
     () => {
       Invoices.createDefaultInvoice(invoice, vendorPrimaryAddress);
-      Invoices.checkCreatedInvoice(invoice, vendorPrimaryAddress);
+      Invoices.createInvoiceLine(invoiceLine);
+      Invoices.checkInvoiceLine(invoiceLine);
       Invoices.deleteInvoiceViaActions();
       Invoices.confirmInvoiceDeletion();
     },
