+++ cypress/e2e/invoices/pol-payment-status-is-changed-to-artially-paid-after-cancelling-one-of-related paid-invoice.cy.js
@@ -22,7 +22,7 @@ import BasicOrderLine from '../../support/fragments/orders/basicOrderLine';
 import MaterialTypes from '../../support/fragments/settings/inventory/materialTypes';
 import OrderLineDetails from '../../support/fragments/orders/orderLineDetails';
 
-describe('Incvoices', () => {
+describe('Invoices', () => {
   const defaultFiscalYear = { ...FiscalYears.defaultUiFiscalYear };
   const defaultLedger = { ...Ledgers.defaultUiLedger };
   const defaultFund = { ...Funds.defaultUiFund };
